+++ test/Confluent.SchemaRegistry.Serdes.IntegrationTests/Tests/SyncOverAsync.cs
@@ -50,7 +50,7 @@ namespace Confluent.SchemaRegistry.Serdes.IntegrationTests
             
             var schemaRegistryConfig = new SchemaRegistryConfig
             {
-                SchemaRegistryUrl = schemaRegistryServers
+                Url = schemaRegistryServers
             };
 
             var topic = Guid.NewGuid().ToString();